requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
python-dotenv
google-generativeai
python-multipart
pymongo 